lab5_mlflow/pyproject.toml
[project]
name = "lab5-mlflow"
version = "0.1.0"
description = "MLflow lab with Airflow DAGs and FastAPI model serving"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mlflow==3.4",
  "scikit-learn>=1.4,<2",
  "pandas>=2.2,<3",
  "numpy>=2,<3",
  "protobuf>=4,<6",
  "rich<14",
  "fastapi==0.115.0",
  "uvicorn[standard]==0.31.0",
  "python-multipart==0.0.9",
  "pydantic>=2.6,<3",
  "aiofiles>=23.0",
  "starlette>=0.38",
  "snowflake-connector-python>=3.6,<4",
  "snowflake-sqlalchemy>=1.5,<2",
  "neo4j>=5.14,<6",
  "py2neo>=2021.2",
  "python-dotenv>=1.0,<2",
]
